clicker/requirements.txt
fastapi==0.100.1
pydantic[email]==2.8.2
uvicorn==0.29.0
Pillow==10.3.0
minio==7.1.12
python-dotenv==1.0.1
websockets==14.1
requests>=2.32.3
fuzzywuzzy==0.18.0
vertexai==1.71.0
google-auth==2.35.0
instructor==1.7.0
google-genai==0.3.0
celery==5.4.0
aio_pika==9.5.4
redis==5.2.1
psycopg2-binary==2.9.9
asyncpg==0.30.0
sqlalchemy==2.0.30
langchain-openai==0.3.4
langchain-anthropic>=0.3.13
langgraph==0.2.70
langfuse==3.1.3
langchain==0.3.19
httpx
Levenshtein
md2tgmd==0.3.9
uncurl==0.0.11